JL8-26 refator: 플레이리스트 곡 담기 리팩토링
